--- controllers/auth.js
@@ -1,157 +1,0 @@
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const gravatar = require("gravatar");
-const fs = require("fs/promises");
-const path = require("path");
-const Jimp = require("jimp");
-const { uid } = require("uid");
-
-const { ctrlWrapper } = require("../utils");
-
-const { User } = require("../models/user");
-
-const { HttpError, sendEmail } = require("../helpers");
-
-const { SECRET_KEY, BASE_URL } = process.env;
-
-const avatarsDir = path.join(__dirname, "../", "public", "avatars");
-
-const register = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  if (user) {
-    throw HttpError(409, `Email ${email} alredy in use`);
-  }
-
-  const hashPassword = await bcrypt.hash(password, 10);
-  const avatarURL = gravatar.url(email);
-  const verificationToken = uid();
-
-  const result = await User.create({
-    ...req.body,
-    password: hashPassword,
-    avatarURL,
-    verificationToken,
-  });
-  if (!result) {
-    throw HttpError(404, `Creation failed`);
-  }
-
-  const verifyEmail = {
-    to: email,
-    subject: "Verify Email",
-    html: `<a target="_blank" href="${BASE_URL}/api/auth/verify/${verificationToken}">Click verify email</a>`,
-  };
-  await sendEmail(verifyEmail);
-
-  res.status(201).json({
-    name: result.name,
-    email: result.email,
-  });
-};
-
-const verifyEmail = async (req, res) => {
-  const { verificationToken } = req.params;
-  const user = await User.findOne({ verificationToken });
-  if (!user) {
-    throw HttpError(404, "User not found");
-  }
-  await User.findByIdAndUpdate(user._id, {
-    verify: true,
-    verificationToken: null,
-  });
-  res.json({ message: "Verification successful" });
-};
-
-const resendVerifyEmail = async (req, res) => {
-  const { email } = req.body;
-  const user = await User.findOne({ email });
-  if (!user) {
-    throw HttpError(400, "missing required field email");
-  }
-  if (user.verify) {
-    throw HttpError(400, "Verification has already been passed");
-  }
-
-  const verifyEmail = {
-    to: email,
-    subject: "Verify Email",
-    html: `<a target="_blank" href="${BASE_URL}/api/auth/verify/${user.verificationToken}">Click verify email</a>`,
-  };
-  await sendEmail(verifyEmail);
-  res.json({ message: "Verification email sent" });
-};
-
-const login = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  if (!user) {
-    throw HttpError(401, "Email or password is wrong");
-  }
-  if (!user.verify) {
-    throw HttpError(401, "Email not verify");
-  }
-  const passwordCompare = await bcrypt.compare(password, user.password);
-  if (!passwordCompare) {
-    throw HttpError(401, "Email or password is wrong");
-  }
-
-  const payload = {
-    id: user._id,
-  };
-
-  const token = jwt.sign(payload, SECRET_KEY, { expiresIn: "23h" });
-
-  await User.findByIdAndUpdate(user._id, { token });
-
-  res.json({
-    token,
-  });
-};
-
-const getCurrent = async (req, res) => {
-  const { name, email } = req.user;
-
-  res.json({
-    name,
-    email,
-  });
-};
-
-const logout = async (req, res) => {
-  const { _id } = req.user;
-  await User.findByIdAndUpdate(_id, { token: "" });
-
-  res.json({
-    message: "Logout success",
-  });
-};
-
-const updateAvatar = async (req, res) => {
-  const { _id } = req.user;
-  const { path: tempUpload, filename } = req.file;
-  const avatarName = `${_id}_${filename}`;
-  const resultUpload = path.join(avatarsDir, avatarName);
-
-  await Jimp.read(resultUpload)
-    .then((resultUpload) => {
-      return resultUpload.resize(250, 250).write(resultUpload);
-    })
-    .catch((error) => console.log(error));
-
-  await fs.rename(tempUpload, resultUpload);
-  const avatarURL = path.join("avatars", avatarName);
-  await User.findByIdAndUpdate(_id, { avatarURL });
-
-  res.json({ avatarURL });
-};
-
-module.exports = {
-  register: ctrlWrapper(register),
-  verifyEmail: ctrlWrapper(verifyEmail),
-  resendVerifyEmail: ctrlWrapper(resendVerifyEmail),
-  login: ctrlWrapper(login),
-  getCurrent: ctrlWrapper(getCurrent),
-  logout: ctrlWrapper(logout),
-  updateAvatar: ctrlWrapper(updateAvatar),
-};
