+++ controllers/auth/index.js
@@ -1,0 +1,17 @@
+const { register } = require("./register");
+const { verifyEmail } = require("./verifyEmail");
+const { resendVerifyEmail } = require("./resendVerifyEmail");
+const { login } = require("./login");
+const { getCurrent } = require("./getCurrent");
+const { logout } = require("./logout");
+const { updateAvatar } = require("./updateAvatar");
+
+module.exports = {
+  register,
+  verifyEmail,
+  resendVerifyEmail,
+  login,
+  getCurrent,
+  logout,
+  updateAvatar,
+};
