+++ controllers/auth/logout.js
@@ -1,0 +1,14 @@
+const { ctrlWrapper } = require("../../utils");
+
+const { User } = require("../../models/user");
+
+const logout = async (req, res) => {
+  const { _id } = req.user;
+  await User.findByIdAndUpdate(_id, { token: "" });
+
+  res.json({
+    message: "Logout success",
+  });
+};
+
+module.exports = { logout: ctrlWrapper(logout) };
