+++ controllers/auth/verifyEmail.js
@@ -1,0 +1,22 @@
+const { ctrlWrapper } = require("../../utils");
+
+const { User } = require("../../models/user");
+
+const { HttpError } = require("../../helpers");
+
+const verifyEmail = async (req, res) => {
+  const { verificationToken } = req.params;
+  const user = await User.findOne({ verificationToken });
+  if (!user) {
+    throw HttpError(404, "User not found");
+  }
+  await User.findByIdAndUpdate(user._id, {
+    verify: true,
+    verificationToken: null,
+  });
+  res.json({ message: "Verification successful" });
+};
+
+module.exports = {
+  verifyEmail: ctrlWrapper(verifyEmail),
+};
