+++ controllers/contacts/index.js
@@ -1,0 +1,15 @@
+const { listContacts } = require("./listContacts");
+const { getContactById } = require("./getContactById");
+const { addContact } = require("./addContact");
+const { removeContact } = require("./removeContact");
+const { updateById } = require("./updateById");
+const { updateStatusContact } = require("./updateStatusContact");
+
+module.exports = {
+  listContacts,
+  getContactById,
+  addContact,
+  removeContact,
+  updateById,
+  updateStatusContact,
+};
