+++ helpers/index.js
@@ -1,7 +1,7 @@
 const HttpError = require("./HttpError");
-const sendEmail = require("./sendEmail");
+// const sendEmail = require("../services/email/sendEmail");
 
 module.exports = {
   HttpError,
-  sendEmail,
+  // sendEmail,
 };
